+++ compiler/src/test/java/com/squareup/anvil/compiler/dagger/ProvidesMethodFactoryGeneratorTest.kt
@@ -1,7 +1,6 @@
 package com.squareup.anvil.compiler.dagger
 
 import com.google.common.truth.Truth.assertThat
-import com.squareup.anvil.compiler.USE_IR
 import com.squareup.anvil.compiler.WARNINGS_AS_ERRORS
 import com.squareup.anvil.compiler.anvilModule
 import com.squareup.anvil.compiler.componentInterface
@@ -3510,7 +3509,6 @@ public final class DaggerModule1_ProvideFunctionFactory implements Factory<Set<S
     sources = sources,
     enableDaggerAnnotationProcessor = enableDagger,
     generateDaggerFactories = !enableDagger,
-    useIR = USE_IR,
     allWarningsAsErrors = WARNINGS_AS_ERRORS,
     block = block,
     previousCompilationResult = previousCompilationResult,