+++ CHANGELOG.md
@@ -15,9 +15,14 @@
 
 ### Fixed
 
+- Fix duplicate generated binding method names. If a class contributes multiple bindings and the bound types have the same short name, then Anvil would generate methods with duplicate names that clash in the end.
+- Support `Any` as bound type when using `@ContributesBinding` or `@ContributesMultibinding`, see #680.
+
 ### Security
 
 ### Custom Code Generator
+
+- Add option to change the JVM target when testing code generators with the custom `AnvilCompilation` class, see #682.
 
 # [2.4.4] - 2023-01-12